Refactor tick engine to float TM and explicit gate threshold

+++ rsl_turn_sequencing/models.py
@@ -4,5 +4,6 @@ from dataclasses import dataclass
 @dataclass
 class Actor:
     name: str
-    speed: int
-    turn_meter: int = 0
+    # Speed and turn meter are modeled as floating point values.
+    speed: float
+    turn_meter: float = 0.0
